--- article-extractor.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://linkmove.io/xsd/extractor_config_2.xsd"
-        xmlns="http://linkmove.io/xsd/extractor_config_2.xsd">
-    <type>csv</type>
-    <connectorId>articleSourceConnector</connectorId>
-    <extractor>
-        <attributes>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>id</target>
-            </attribute>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>body</target>
-            </attribute>
-            <attribute>
-                <type>java.util.Date</type>
-                <target>publishedOn</target>
-            </attribute>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>title</target>
-            </attribute>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>domain</target>
-            </attribute>
-        </attributes>
-        <properties>
-            <extractor.csv.delimiter>;</extractor.csv.delimiter>
-            <extractor.csv.readFrom>2</extractor.csv.readFrom>
-        </properties>
-    </extractor>
-</config>
