--- config.yml
@@ -1,16 +1,0 @@
-jdbc:
-  derby:
-    url: jdbc:derby:target/demodb;create=true
-    driverClassName: org.apache.derby.jdbc.EmbeddedDriver
-    initialSize: 1
-
-cayenne:
-  datasource: derby #target database
-  createSchema: true
-
-linkmove:
-  connectorFactories:
-    - type: uri
-      connectors: #data sources
-        domainSourceConnector: file:///Users/user/bootique-linkmove-demo/domain.json # use an absolute URI
-        tagSourceConnector: file:///Users/user/bootique-linkmove-demo/tag.csv
