--- src/main/java/io/bootique/lm/cayenne/Domain.java
@@ -1,9 +1,0 @@
-package io.bootique.lm.cayenne;
-
-
-import io.bootique.lm.cayenne.auto._Domain;
-
-public class Domain extends _Domain {
-
-    private static final long serialVersionUID = 1L;
-}
