--- src/main/java/io/bootique/lm/cayenne/Tag.java
@@ -1,10 +1,0 @@
-package io.bootique.lm.cayenne;
-
-
-import io.bootique.lm.cayenne.auto._Tag;
-
-public class Tag extends _Tag {
-
-    private static final long serialVersionUID = 1L; 
-
-}
