--- src/main/java/io/bootique/lm/cayenne/auto/_Tag.java
@@ -1,39 +1,0 @@
-package io.bootique.lm.cayenne.auto;
-
-import io.bootique.lm.cayenne.Article;
-import org.apache.cayenne.CayenneDataObject;
-import org.apache.cayenne.exp.Property;
-
-/**
- * Class _Tag was generated by Cayenne.
- * It is probably a good idea to avoid changing this class manually,
- * since it may be overwritten next time code is regenerated.
- * If you need to make any customizations, please use subclass.
- */
-public abstract class _Tag extends CayenneDataObject {
-
-    private static final long serialVersionUID = 1L;
-
-    public static final String ID_PK_COLUMN = "id";
-
-    public static final Property<String> NAME = Property.create("name", String.class);
-    public static final Property<Article> ARTICLE = Property.create("article", Article.class);
-
-    public void setName(String name) {
-        writeProperty("name", name);
-    }
-
-    public String getName() {
-        return (String) readProperty("name");
-    }
-
-    public void setArticle(Article article) {
-        setToOneTarget("article", article, true);
-    }
-
-    public Article getArticle() {
-        return (Article) readProperty("article");
-    }
-
-
-}
