--- src/main/resources/cayenne-project.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<domain project-version="9">
-    <map name="datamap"/>
-    <!--No data node is declared - config.yml is used-->
-</domain>
