--- src/main/resources/datamap.map.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<data-map xmlns="http://cayenne.apache.org/schema/9/modelMap"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://cayenne.apache.org/schema/9/modelMap http://cayenne.apache.org/schema/9/modelMap.xsd"
-          project-version="9">
-    <property name="defaultPackage" value="io.bootique.linkmove.demo"/>
-    <property name="quoteSqlIdentifiers" value="true"/>
-    <db-entity name="article">
-        <db-attribute name="body" type="LONGVARCHAR" length="16777215"/>
-        <db-attribute name="domain_id" type="BIGINT" isMandatory="true" length="19"/>
-        <db-attribute name="id" type="BIGINT" isPrimaryKey="true" isGenerated="true" isMandatory="true" length="19"/>
-        <db-attribute name="published_on" type="TIMESTAMP" isMandatory="true"/>
-        <db-attribute name="title" type="VARCHAR" isMandatory="true" length="100"/>
-    </db-entity>
-    <db-entity name="domain">
-        <db-attribute name="id" type="BIGINT" isPrimaryKey="true" isGenerated="true" isMandatory="true" length="19"/>
-        <db-attribute name="name" type="VARCHAR" isMandatory="true" length="100"/>
-        <db-attribute name="vhost" type="VARCHAR" isMandatory="true" length="100"/>
-    </db-entity>
-    <db-entity name="tag">
-        <db-attribute name="article_id" type="BIGINT" isMandatory="true"/>
-        <db-attribute name="id" type="BIGINT" isPrimaryKey="true" isGenerated="true" isMandatory="true"/>
-        <db-attribute name="name" type="VARCHAR" length="200"/>
-    </db-entity>
-    <obj-entity name="Article" className="io.bootique.lm.cayenne.Article" dbEntityName="article">
-        <obj-attribute name="body" type="java.lang.String" db-attribute-path="body"/>
-        <obj-attribute name="publishedOn" type="java.util.Date" db-attribute-path="published_on"/>
-        <obj-attribute name="title" type="java.lang.String" db-attribute-path="title"/>
-        <post-add method-name="onPostAdd"/>
-    </obj-entity>
-    <obj-entity name="Domain" className="io.bootique.lm.cayenne.Domain" dbEntityName="domain">
-        <obj-attribute name="name" type="java.lang.String" db-attribute-path="name"/>
-        <obj-attribute name="vhost" type="java.lang.String" db-attribute-path="vhost"/>
-    </obj-entity>
-    <obj-entity name="Tag" className="io.bootique.lm.cayenne.Tag" dbEntityName="tag">
-        <obj-attribute name="name" type="java.lang.String" db-attribute-path="name"/>
-    </obj-entity>
-    <db-relationship name="domain" source="article" target="domain" toMany="false">
-        <db-attribute-pair source="domain_id" target="id"/>
-    </db-relationship>
-    <db-relationship name="tags" source="article" target="tag" toMany="true">
-        <db-attribute-pair source="id" target="article_id"/>
-    </db-relationship>
-    <db-relationship name="articles" source="domain" target="article" toMany="true">
-        <db-attribute-pair source="id" target="domain_id"/>
-    </db-relationship>
-    <db-relationship name="article" source="tag" target="article" toMany="false">
-        <db-attribute-pair source="article_id" target="id"/>
-    </db-relationship>
-    <obj-relationship name="domain" source="Article" target="Domain" deleteRule="Nullify" db-relationship-path="domain"/>
-    <obj-relationship name="tags" source="Article" target="Tag" deleteRule="Deny" db-relationship-path="tags"/>
-    <obj-relationship name="articles" source="Domain" target="Article" deleteRule="Deny" db-relationship-path="articles"/>
-    <obj-relationship name="article" source="Tag" target="Article" deleteRule="Nullify" db-relationship-path="article"/>
-</data-map>
