--- src/main/resources/etl/article.csv
@@ -1,4 +1,0 @@
-Id;Body;Published on;Title;Domain
-1;ObjectStyle has deep roots in the open source community...;;About us;1
-2;Getting Started with Bootique...;;Getting Started with Bootique;2
-3;LinkMove  an ETL solution...;;LinkMove;3
