+++ sync-database/article-extractor.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://linkmove.io/xsd/extractor_config_2.xsd"
+        xmlns="http://linkmove.io/xsd/extractor_config_2.xsd">
+    <type>jdbc</type>
+    <connectorId>sourcedb</connectorId>
+    <extractor>
+        <attributes>
+            <attribute>
+                <type>java.lang.String</type>
+                <source>id</source>
+                <target>db:id</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <source>body</source>
+                <target>body</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <source>title</source>
+                <target>title</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <source>domain_id</source>
+                <target>db:domain_id</target>
+            </attribute>
+        </attributes>
+        <properties>
+            <!--
+             Query to run against the source. Supports full Cayenne
+                             SQLTemplate syntax, including parameters and directives.
+            -->
+            <extractor.jdbc.sqltemplate>SELECT id, title, body, domain_id FROM s_article</extractor.jdbc.sqltemplate>
+            <extractor.jdbc.sqltemplate.caps>LOWER</extractor.jdbc.sqltemplate.caps>
+        </properties>
+    </extractor>
+</config>
