+++ sync-database/config.yml
@@ -1,0 +1,22 @@
+jdbc:
+  sourcedb:
+    url: jdbc:mysql://localhost:3306/sourcedb?connectTimeout=0&autoReconnect=true
+    driverClassName: com.mysql.jdbc.Driver
+    initialSize: 1
+    username: root
+    password:
+  targetdb:
+      url: jdbc:mysql://localhost:3306/targetdb?connectTimeout=0&autoReconnect=true
+      driverClassName: com.mysql.jdbc.Driver
+      initialSize: 1
+      username: root
+      password:
+
+cayenne:
+  datasource: targetdb
+  createSchema: true
+
+linkmove:
+  extractorsDir: /Users/your_user/bootique-linkmove-demo/sync-database #use absolute path
+  connectorFactories:
+    - type: jdbc
