+++ sync-database/src/main/java/io/bootique/lm/cayenne/TArticle.java
@@ -1,0 +1,9 @@
+package io.bootique.lm.cayenne;
+
+import io.bootique.lm.cayenne.auto._TArticle;
+
+public class TArticle extends _TArticle {
+
+    private static final long serialVersionUID = 1L; 
+
+}
