+++ sync-database/src/main/java/io/bootique/lm/cayenne/TDomain.java
@@ -1,0 +1,9 @@
+package io.bootique.lm.cayenne;
+
+import io.bootique.lm.cayenne.auto._TDomain;
+
+public class TDomain extends _TDomain {
+
+    private static final long serialVersionUID = 1L; 
+
+}
