+++ sync-database/src/main/java/io/bootique/lm/cayenne/TTag.java
@@ -1,0 +1,9 @@
+package io.bootique.lm.cayenne;
+
+import io.bootique.lm.cayenne.auto._TTag;
+
+public class TTag extends _TTag {
+
+    private static final long serialVersionUID = 1L; 
+
+}
