+++ sync-database/src/main/java/io/bootique/lm/cayenne/auto/_TDomain.java
@@ -1,0 +1,51 @@
+package io.bootique.lm.cayenne.auto;
+
+import io.bootique.lm.cayenne.TArticle;
+import org.apache.cayenne.CayenneDataObject;
+import org.apache.cayenne.exp.Property;
+
+import java.util.List;
+
+/**
+ * Class _TDomain was generated by Cayenne.
+ * It is probably a good idea to avoid changing this class manually,
+ * since it may be overwritten next time code is regenerated.
+ * If you need to make any customizations, please use subclass.
+ */
+public abstract class _TDomain extends CayenneDataObject {
+
+    private static final long serialVersionUID = 1L; 
+
+    public static final String ID_PK_COLUMN = "id";
+
+    public static final Property<String> NAME = Property.create("name", String.class);
+    public static final Property<String> VHOST = Property.create("vhost", String.class);
+    public static final Property<List<TArticle>> T_ARTICLES = Property.create("tArticles", List.class);
+
+    public void setName(String name) {
+        writeProperty("name", name);
+    }
+    public String getName() {
+        return (String)readProperty("name");
+    }
+
+    public void setVhost(String vhost) {
+        writeProperty("vhost", vhost);
+    }
+    public String getVhost() {
+        return (String)readProperty("vhost");
+    }
+
+    public void addToTArticles(TArticle obj) {
+        addToManyTarget("tArticles", obj, true);
+    }
+    public void removeFromTArticles(TArticle obj) {
+        removeToManyTarget("tArticles", obj, true);
+    }
+    @SuppressWarnings("unchecked")
+    public List<TArticle> getTArticles() {
+        return (List<TArticle>)readProperty("tArticles");
+    }
+
+
+}
