+++ sync-database/src/main/resources/cayenne-project.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<domain project-version="9">
+	<map name="datamap"/>
+
+	<!--skip data node declaration - config.yml is used-->
+
+	<!--<node name="datanode"-->
+		 <!--factory="org.apache.cayenne.configuration.server.XMLPoolingDataSourceFactory"-->
+		 <!--schema-update-strategy="org.apache.cayenne.access.dbsync.SkipSchemaUpdateStrategy"-->
+		<!--&gt;-->
+		<!--<map-ref name="datamap"/>-->
+		<!--<data-source>-->
+			<!--<driver value="com.mysql.jdbc.Driver"/>-->
+			<!--<url value="jdbc:mysql://localhost:3306/targetdb?connectTimeout=0&amp;autoReconnect=true"/>-->
+			<!--<connectionPool min="1" max="1"/>-->
+			<!--<login userName="root"/>-->
+		<!--</data-source>-->
+	<!--</node>-->
+</domain>
