+++ sync-files-database/article-extractor.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://linkmove.io/xsd/extractor_config_2.xsd"
+        xmlns="http://linkmove.io/xsd/extractor_config_2.xsd">
+    <type>csv</type>
+    <connectorId>articleSourceConnector</connectorId>
+    <extractor>
+        <attributes>
+            <attribute>
+                <type>java.lang.String</type>
+                <target>db:id</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <target>body</target>
+            </attribute>
+            <attribute>
+                <type>java.util.Date</type>
+                <target>db:published_on</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <target>title</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <target>db:domain_id</target>
+            </attribute>
+            <attribute>
+                <type>java.lang.String</type>
+                <target>db:quote_index</target>
+            </attribute>
+        </attributes>
+        <properties>
+            <extractor.csv.delimiter>;</extractor.csv.delimiter>
+            <extractor.csv.readFrom>2</extractor.csv.readFrom>
+        </properties>
+    </extractor>
+</config>
