+++ sync-files-database/config.yml
@@ -1,0 +1,18 @@
+jdbc:
+  targetdb:
+      url: jdbc:mysql://localhost:3306/targetdb?connectTimeout=0&autoReconnect=true
+      driverClassName: com.mysql.jdbc.Driver
+      initialSize: 1
+      username: root
+      password:
+
+cayenne:
+  datasource: targetdb
+
+linkmove:
+  extractorsDir: /Users/your_user/bootique-linkmove-demo/sync-files-database
+  connectorFactories:
+    - type: uri
+      connectors: #sources
+        tagSourceConnector: file:///Users/your_user/bootique-linkmove-demo/sync-files-database/tag.csv #use absolute path
+
