+++ sync-files-database/src/main/java/io/bootique/lm/cayenne/auto/_TTag.java
@@ -1,0 +1,39 @@
+package io.bootique.lm.cayenne.auto;
+
+import org.apache.cayenne.CayenneDataObject;
+import org.apache.cayenne.exp.Property;
+
+import io.bootique.lm.cayenne.TArticle;
+
+/**
+ * Class _TTag was generated by Cayenne.
+ * It is probably a good idea to avoid changing this class manually,
+ * since it may be overwritten next time code is regenerated.
+ * If you need to make any customizations, please use subclass.
+ */
+public abstract class _TTag extends CayenneDataObject {
+
+    private static final long serialVersionUID = 1L; 
+
+    public static final String ID_PK_COLUMN = "id";
+
+    public static final Property<String> NAME = Property.create("name", String.class);
+    public static final Property<TArticle> ARTICLE = Property.create("article", TArticle.class);
+
+    public void setName(String name) {
+        writeProperty("name", name);
+    }
+    public String getName() {
+        return (String)readProperty("name");
+    }
+
+    public void setArticle(TArticle article) {
+        setToOneTarget("article", article, true);
+    }
+
+    public TArticle getArticle() {
+        return (TArticle)readProperty("article");
+    }
+
+
+}
