+++ sync-files-database/src/main/resources/etl/domain.json
