+++ sync-files-database/tag.csv
@@ -1,0 +1,5 @@
+Id;Name;Article
+Id;Name;Article
+1;ObjectStyle;1
+2;Bootique;2
+3;LinkMove;3
