--- tag-extractor.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://linkmove.io/xsd/extractor_config_2.xsd"
-        xmlns="http://linkmove.io/xsd/extractor_config_2.xsd">
-    <type>csv</type>
-    <connectorId>tagSourceConnector</connectorId>
-    <extractor>
-        <attributes>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>id</target>
-            </attribute>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>name</target>
-            </attribute>
-            <attribute>
-                <type>java.lang.String</type>
-                <target>article</target>
-            </attribute>
-        </attributes>
-        <properties>
-            <extractor.csv.delimiter>;</extractor.csv.delimiter>
-            <extractor.csv.readFrom>3</extractor.csv.readFrom>
-        </properties>
-    </extractor>
-</config>
