--- tag.csv
@@ -1,5 +1,0 @@
-Skipped;Skipped;Skipped
-Id;Name;Article
-1;ObjectStyle;1
-2;Bootique;2
-3;LinkMove;3
